toolkit/devtools/server/tests/mochitest/test_inspector-resolve-url.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=921102
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 921102</title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
Components.utils.import("resource://gre/modules/devtools/Loader.jsm");
const {Promise: promise} = Components.utils.import("resource://gre/modules/Promise.jsm", {});
const inspector = devtools.require("devtools/server/actors/inspector");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

var gInspector;
var gDoc;

addTest(function() {
  let url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gDoc = doc;
    let {InspectorFront} = devtools.require("devtools/server/actors/inspector");
    gInspector = InspectorFront(client, tab);
    runNextTest();
  });
});

addTest(function() {
  info("Resolve a relative URL without providing a context node");
  gInspector.resolveRelativeURL("test.png?id=4#wow").then(url => {
    is(url, "chrome://mochitests/content/chrome/toolkit/devtools/server/tests/" +
            "mochitest/test.png?id=4#wow");
    runNextTest();
  });
});

addTest(function() {
  info("Resolve an absolute URL without providing a context node");
  gInspector.resolveRelativeURL("chrome://mochitests/content/chrome/toolkit/" +
                                "devtools/server/").then(url => {
    is(url, "chrome://mochitests/content/chrome/toolkit/devtools/server/");
    runNextTest();
  });
});

addTest(function() {
  info("Resolve a relative URL providing a context node");
  let node = gDoc.querySelector(".big-horizontal");
  gInspector.resolveRelativeURL("test.png?id=4#wow", node).then(url => {
    is(url, "chrome://mochitests/content/chrome/toolkit/devtools/server/tests/" +
            "mochitest/test.png?id=4#wow");
    runNextTest();
  });
});

addTest(function() {
  info("Resolve an absolute URL providing a context node");
  let node = gDoc.querySelector(".big-horizontal");
  gInspector.resolveRelativeURL("chrome://mochitests/content/chrome/toolkit/" +
                                "devtools/server/", node).then(url => {
    is(url, "chrome://mochitests/content/chrome/toolkit/devtools/server/");
    runNextTest();
  });
});

addTest(function() {
  gInspector = gDoc = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=921102">Mozilla Bug 921102</a>
<a id="inspectorContent" target="_blank" href="inspector_getImageData.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
